refatoracao da funcao parametrosParaBusca
