+++ src/util/criaDataFim.js
@@ -1,0 +1,7 @@
+function criaDataFim() {
+  const dataFim = new Date();
+  dataFim.setUTCMonth(dataFim.getUTCMonth() + 1);
+  dataFim.setUTCDate(0);
+  return `${dataFim.getFullYear()}-${dataFim.getUTCMonth() + 1}-${dataFim.getUTCDate()}`;
+}
+module.exports = criaDataFim;
